# aphrodite-backend/docs/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "aphrodite",
        "description": "API Documentation",
        "version": "unknown"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8083/aphrodite-web-api",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
        "/api/v1/users/user/{username}": {
            "get": {
                "tags": ["User Management"],
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "put": {
                "tags": ["User Management"],
                "operationId": "createUpdateUser",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AphroditeNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "aphrodite"
                    }
                }
            ]
        },
        "/api/v1/roles/{rolesName}": {
            "get": {
                "tags": ["Role Management"],
                "operationId": "getRole",
                "parameters": [
                    {
                        "name": "rolesName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "put": {
                "tags": ["Role Management"],
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "rolesName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRoleRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "delete": {
                "tags": ["Role Management"],
                "operationId": "deleteRole",
                "parameters": [
                    {
                        "name": "rolesName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AphroditeNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "aphrodite"
                    }
                }
            ]
        },
...
        "/api/v1/token/refresh": {
            "post": {
                "tags": ["Token Management"],
                "summary": "Refresh token",
                "operationId": "refreshToken",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApiKeyRefreshRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiKey"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AphroditeNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "aphrodite"
                    }
                }
            ]
        },
        "/api/v1/token/invalidate": {
            "post": {
                "tags": ["Token Management"],
                "summary": "Invalidate token",
                "operationId": "invalidateToken",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AphroditeNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "aphrodite"
                    }
                }
            ]
        },
...
        "/api/v1/token/validate": {
            "get": {
                "tags": ["Token Management"],
                "summary": "Validate token",
                "operationId": "validateToken",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiKey"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AphroditeNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "aphrodite"
                    }
                }
            ]
        },
...
    },
    "components": {
        "schemas": {
            "UpdateUserRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "locked": {
                        "type": "boolean"
                    },
                    "enabled": {
                        "type": "boolean"
                    }
                }
            },
            "Namespace": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "Privilege": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "Role": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "privileges": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Privilege"
                        }
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "roles": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    },
                    "expired": {
                        "type": "boolean",
                        "writeOnly": true
                    },
                    "locked": {
                        "type": "boolean",
                        "writeOnly": true
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "namespace": {
                        "$ref": "#/components/schemas/Namespace"
                    }
                }
            },
            "UpdateRoleRequest": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "privileges": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "ApiKeyRefreshRequest": {
                "type": "object",
                "properties": {
                    "validitySeconds": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "ApiKey": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "lastUsedAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "expireAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "originIp": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "namespace": {
                        "$ref": "#/components/schemas/Namespace"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Aphrodite-API-Key": {
                "type": "apiKey",
                "name": "Aphrodite-API-Key",
                "in": "header"
            }
        }
    }
}